pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "paperscout"
version = "0.1.0"
description = "Agentic AI research assistant that discovers, evaluates, and reports on academic papers"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langchain-groq>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "pymupdf>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "arize-phoenix>=8.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
]

[project.scripts]
paperscout = "paperscout.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
